+++ src/put/ft_putnbr_base.c
@@ -6,13 +6,13 @@
 /*   By: msloot <msloot@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/27 10:55:10 by msloot            #+#    #+#             */
-/*   Updated: 2023/11/18 18:11:30 by msloot           ###   ########.fr       */
+/*   Updated: 2023/11/18 18:45:46 by msloot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-ssize_t	ft_putnbr_base(long n, const char *base)
+ssize_t	ft_putnbr_base(ssize_t n, const char *base)
 {
 	char	str[12];
 	size_t	base_len;
